requirements.txt
# Core
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
aiohttp>=3.9.0

# Local Models
ollama>=0.1.0

# Voice
openai-whisper>=20231117
pyttsx3>=2.90
pyaudio>=0.2.14

# CLI
click>=8.1.0
rich>=13.0.0
